fix: 박람회 생성 ispending 적용
