+++ src/shared/types/exhibition/create/type.ts
@@ -1,32 +1,10 @@
-export type ExhibitionFormData = {
-  title: string;
-  introduction: string;
-  address: string;
-  location: string;
-  trainings: {
-    id?: number;
-    title: string;
-    startedAt: string;
-    endedAt: string;
-    category?: 'ESSENTIAL' | 'CHOICE';
-  }[];
-  standard: {
-    id?: number;
-    title: string;
-    startedAt: string;
-    endedAt: string;
-  }[];
-  image: File | null | string;
-  startedDay: string;
-  finishedDay: string;
-};
-
 import {
   UseFieldArrayReturn,
   UseFormRegister,
   UseFormSetValue,
   UseFormWatch,
 } from 'react-hook-form';
+import { ExhibitionFormData } from '../type';
 
 export interface CreateExhibitionData {
   title: string;
