+++ src/shared/ui/CheckBox/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { UseFormSetValue, UseFormWatch } from 'react-hook-form';
 import { Check } from '@/shared/assets/icons';
-import { ExhibitionFormData } from '@/shared/types/exhibition/create/type';
+import { ExhibitionFormData } from '@/shared/types/exhibition/type';
 
 interface CheckBoxProps {
   text: string;
