+++ src/views/exhibition/create/ui/createExhibition/index.tsx
@@ -6,12 +6,13 @@ import ExhibitionForm from '@/widgets/exhibition/ui/ExhibitionForm';
 import { Header } from '@/widgets/layout';
 
 const CreateExhibition = () => {
-  const mutation = useCreateExhibitionMutation();
+  const { mutate, isPending, isSuccess } = useCreateExhibitionMutation();
+
   return (
     <div className="flex min-h-screen flex-col gap-[30px]">
       <Header />
       <div className="flex flex-1 justify-center p-16">
-        <ExhibitionForm mutation={mutation} />
+        <ExhibitionForm mutation={{ mutate, isPending, isSuccess }} />
       </div>
     </div>
   );
