+++ src/widgets/exhibition/edit/model/handleEditExhibitionFormSubmit.ts
@@ -2,7 +2,7 @@ import { QueryClient } from '@tanstack/react-query';
 import { useRouter } from 'next/navigation';
 import { toast } from 'react-toastify';
 import { generateUniqueId } from '@/entities/exhibition/model/generateUniqueId';
-import { ExhibitionFormData } from '@/shared/types/exhibition/create/type';
+import { ExhibitionFormData } from '@/shared/types/exhibition/type';
 import { convertAddressToCoordinates } from '../../api/convertAddressToCoordinates';
 import { uploadImage } from '../../api/uploadImage';
 import { editExhibition } from '../api/editExhibition';
